requirements.txt
flask
openai
groq
python-dotenv
